issue-114-리스트 결정화면 단계 축소 / 디테일,결정화면 UI수정
