+++ app/src/main/res/menu/menu_create_list.xml
@@ -2,6 +2,14 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
+
+    <item
+        android:id="@+id/goods_search"
+        android:icon="@drawable/btn_search"
+        android:title="@string/menu_search"
+        app:actionViewClass="androidx.appcompat.widget.SearchView"
+        app:showAsAction="always" />
+
     <item
         android:id="@+id/btn_show_cart"
         android:icon="@drawable/btn_shopping_cart"
