--- app/src/main/res/menu/menu_save_goods.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/btn_decide"
-        app:showAsAction="ifRoom"
-        android:icon="@drawable/btn_all_confirm"
-        android:title="@string/decide" />
-</menu>
